User_Details/requirements.txt
flask-ngrok==0.0.25
pyjwt==1.7.1
alembic==1.4.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
flask-login
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
stripe==2.54.0
urllib3==1.25.10
Werkzeug==1.0.1
